Fix a bug that casts float value to int during adding an ad

+++ src/Entity/Ad.php
@@ -55,7 +55,7 @@ class Ad
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
